fix(web): move fonts to separate module
