+++ apps/web/src/app/fonts.ts
@@ -1,0 +1,8 @@
+import { Lato } from 'next/font/google';
+
+export const lato = Lato({
+  weight: ['300', '400', '700'],
+  display: 'swap',
+  subsets: ['latin'],
+  preload: true,
+});
